Simplify the generation of class files, change json parser and serializer and list handling.
Create new directory "objects_common".
In this directory, there are two python files, arrayType.py and
jsonObject.py. The common objects are imported in each generated class
definition, and the objects_common directory is copied into the output directory of the generator.

arrayType.py:
This defines the ArrayType class that inherits from the python list type.
It stores the class of the items that are placed in the list in its klass
attribute. It has a method to create a new object and place it in the
list (append_new). Its serialize json method builds a Json list,
serializes the items in the list and fills the json list.

jsonObject.py:
The class JsonObject that is defined here is used as the superclass for
any generated object. This class has methods for loading and serializing
json that work on any object. This means the objects do not need to
include a json serializer in the generated code, which greatly simplifies
the code generation.
The list _child_objects contains all the attributes of the object, thus
generating a list of objects is not necessary, which simplifies code
generation. If you do not want to include an attribute in the
_child_objects list, prefix the attribute with underscore, for example
_test. To rebuild the list after adding an attribute call the method
build_child_objects_list. It is important that the superclass constructor
super(ExampleClass, self).__init__(json_string) is called *after* the
attributes have been defined in the __init__ function of an object. This is implemented in the generator.

The json parser (load_json) has been modified.
Objects in Json become dicts in python after using the json.loads() method.
Due to this an incoming dict is assumed to be an object, and the objects load_json method is called.
As a consequence, lists in Json must be translated to lists in python.
Lists that have a key must be lists in python, too.
Lists must not be translated to dicts, otherwise the parser assumes that they are objects.
Lists of lists [[],[],..] can not be parsed in the current state of the
parser.
Handling of lists is simplified by using the arrayType class for lists.
Lists can either contain base types, objects or dicts, the parser can
handle each case.
Any received element that is neither a list nor a dictionary is assumed to
be a base type such as a string, and directly sets an attribute value.

The json serializer works as follows. Objects are serialized using their
serialize_json method, which tries to call "serialize_json" on any of its
attributes in the _child_objects list. If the serialize_json method is not
present on an attribute, the attribute is assumed to be of a base type,
and simple enough that it can be added directly to the output.
The arrayType class has a method that serializes lists.

The json exchanged between client and server does no longer need to
include a 'class'item to identify the class of a field. The server can
determine the class of an item from its name.

The generated class definitions have been reduced to the __init__
definition, and the enum class definition.