Convert unicode to phython byte strings. Fix load_json.
Strings that are the result of json.loads are unicode.
They must be encoded as utf-8 to be compatible whith the python attribute
names, and string values, and to allow strict type checking of JSON
received from a client.

Previously the load_json method checked if the class of an attribute is JsonObject, this was wrong, as JsonObject is usually the superclass, and sometimes not
even that. Now it checks if the load_json method is available instead
(duck typing).