Change handling of PUT and POST.
Use PUT only for creating objects, not for updating objects.
When POST is issued on an existing object, update the given keys, but do
not reset the values of the keys to the default value. If a key is not
given its value does not change.
If a list is part of a POST request, overwrite the list completely with
the given list. For inserting and deleting specific elements, the lists
path can be used.

Remove unused functions "tab" and "handleResponse".
Define ArrayTypeError to report wrong types passed to a list.