Add support for lists with keys to code generator. Avoid dicts.
Look for the statement "x-key" in the swagger json file.
Refactor the functions getType and translateClasses.
getType returns a dict and a boolean.
Remove return statements from getType.
Comment out unused code path in getType.
translateClasses appends the dict to the atts list.

Do not use dictionaries to save objects. Always use a class definition of
type JsonObject. That way the methods serialize_json and load_json are
available.