Simplify JsonObject.load_json.
Add a load_json method to all types: to ArrayType, EnumType and
keyedArrayType.

Move type checking into the load_json method of the types.

Use factory method for ArrayType. Call ArrayType.factory(klass) to
instantiate an ArrayType. This improves the append_new function, because
there is only one check for the load_json method in __init__, and not
repeated checks everytime append_new is called.
Make the presence of the serialize_json method optional, depending on
klass. Lists of base types do not need a serialize_json method.