Handle multiple json files. Use argparse. Copy backend files. Add backend example.
Multiple json files can be specified as positional argument. All files are
handled in a loop. This means the .cop directory is not needed anymore.
The port is not read from the json file, but specified as argument.
When the argument -a is specified the demo server is generated, if not the
abse server is generated.